+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This file documents recent notable changes to this project. The format of this
 file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Added `StreamRequestPayload` enum to encapsulate different stream request
+  types (semi-supervised and time series generator) along with their record type
+  and request data.
+  - Updated `send_stream_request`, `receive_stream_request` functions to use
+    `StreamRequestPayload` for sending and receiving requests, removing the need
+    for `NodeType` enum.
+
 ## [0.23.0] - 2025-06-18
 
 ### Changed
@@ -328,6 +339,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Move from giganto
 
+[Unreleased]: https://github.com/aicers/giganto-client/compare/0.23.0...main
 [0.23.0]: https://github.com/aicers/giganto-client/compare/0.22.0...0.23.0
 [0.22.0]: https://github.com/aicers/giganto-client/compare/0.21.0...0.22.0
 [0.21.0]: https://github.com/aicers/giganto-client/compare/0.20.0...0.21.0